Keep maintenance warnings focused on actionable script debt
Service facades are no longer counted as broad value objects, and the OpenSpec terminology CLI is split so the main tool is below the warning threshold while preserving its JSON validation run.

Constraint: Maintenance burn-down must not hide live scanner findings or broaden service method-count exemptions beyond explicit service facades.
Rejected: Marking accepted service warnings as harmless without changing scanner behavior | the live ledger would still count them as warnings.
Confidence: high
Scope-risk: narrow
Directive: Keep service-facade exemptions tied to Service classes and scanner tests; do not blanket-exempt public method counts.
Tested: npm.cmd test -- --watchAll=false --runTestsByPath scripts/__tests__/maintenance-scanner.test.ts scripts/__tests__/maintenance-warning-ledger.test.ts --runInBand; node scripts\qc\validate-maintenance-warning-ledger.mjs; npm.cmd run maintain:scan:gate; npm.cmd run terminology:validate -- --json; npm.cmd exec prettier -- --check openspec/scripts/terminology-tool.ts openspec/scripts/terminology-tool.types.ts openspec/scripts/terminology-tool.output.js; npx.cmd oxfmt --check docs\qc\maintenance-warning-ledger.json scripts\maintenance\scan-maintenance.mjs scripts\__tests__\maintenance-scanner.test.ts; git diff --cached --check
Not-tested: Full verify:qc was not rerun after this narrow maintenance slice; it had passed immediately before the branch work on merged main.
